+++ packages/networks/src/index.spec.ts
@@ -4,25 +4,25 @@
 import type { SubstrateNetwork } from './types';
 
 import { knownGenesis, knownIcon, knownLedger, knownTestnet } from './defaults';
-import filtered, { all, available } from '.';
+import { allNetworks, availableNetworks, selectableNetworks } from '.';
 
-describe('filtered', (): void => {
+describe('selectableNetworks', (): void => {
   it('has the correct starting order', (): void => {
-    expect(filtered.slice(0, 3).map(({ prefix }) => prefix)).toEqual([0, 2, 42]);
+    expect(selectableNetworks.slice(0, 3).map(({ prefix }) => prefix)).toEqual([0, 2, 42]);
   });
 
   it('has no ignored networks', (): void => {
-    expect(available.some(({ isIgnored }) => isIgnored)).toEqual(false);
+    expect(availableNetworks.some(({ isIgnored }) => isIgnored)).toEqual(false);
   });
 
   it('has no reserved networks', (): void => {
-    expect(available.some(({ prefix }) => prefix === 47)).toEqual(false);
+    expect(availableNetworks.some(({ prefix }) => prefix === 47)).toEqual(false);
   });
 
-  it('has all genesis information', (): void => {
+  it('has allNetworks genesis information', (): void => {
     expect(
       Object.entries(knownGenesis).filter(([network, genesisHash]) =>
-        available.some((a) =>
+        availableNetworks.some((a) =>
           a.network === network &&
           genesisHash.some((g, index) => a.genesisHash[index] !== g)
         )
@@ -30,10 +30,10 @@ describe('filtered', (): void => {
     ).toEqual([]);
   });
 
-  it('has all ledger details', (): void => {
+  it('has allNetworks ledger details', (): void => {
     expect(
       Object.entries(knownLedger).filter(([network, slip44]) =>
-        available.some((a) =>
+        availableNetworks.some((a) =>
           a.network === network && (
             a.slip44 !== slip44 ||
             !a.hasLedgerSupport ||
@@ -47,16 +47,16 @@ describe('filtered', (): void => {
   it('has no testnets exposed', (): void => {
     expect(
       Object.keys(knownTestnet).filter((network) =>
-        available.some((a) =>
+        availableNetworks.some((a) =>
           a.network === network
         )
       )
     ).toEqual([]);
   });
 
-  it('has all icons, except for overrides', (): void => {
+  it('has allNetworks icons, except for overrides', (): void => {
     expect(
-      available.filter(({ icon, network }) =>
+      availableNetworks.filter(({ icon, network }) =>
         icon !== 'substrate' &&
         knownIcon[network] !== icon
       )
@@ -67,7 +67,7 @@ describe('filtered', (): void => {
     const dupes: SubstrateNetwork[] = [];
     const uniques: SubstrateNetwork[] = [];
 
-    all.forEach((a): void => {
+    allNetworks.forEach((a): void => {
       if (uniques.some((u) => u.prefix === a.prefix)) {
         dupes.push(a);
       } else {