+++ packages/util/src/is/buffer.ts
@@ -16,5 +16,5 @@
  * ```
  */
 export function isBuffer (value: unknown): value is Buffer {
-  return Buffer.isBuffer(value);
+  return typeof Buffer !== 'undefined' && Buffer.isBuffer(value);
 }