+++ packages/util/src/is/index.ts
@@ -10,6 +10,7 @@ export { isBn } from './bn';
 export { isBuffer } from './buffer';
 export { isBoolean } from './boolean';
 export { isChildClass } from './childClass';
+export { isCompact } from './compact';
 export { isError } from './error';
 export { isFunction } from './function';
 export { isHex } from './hex';